Moved taskComplete() to script.js Moved clearAll() to script.js
