+++ src/script.js
@@ -1,4 +1,4 @@
-window.clearAll = function() {
+export const clearAll = () => {
 	let taskItems = Array.from(JSON.parse(localStorage.getItem("listItems")));
 	const newList = taskItems.filter(item => {
 		return item.completed !== true;
@@ -7,4 +7,16 @@ window.clearAll = function() {
 	window.location.reload;
 }
 
-export default clearAll;
+
+window.taskCompleted = function(event) {
+	const taskItems = Array.from(JSON.parse(localStorage.getItem("listItems")));
+	taskItems.forEach((item) => {
+		if (item.description === event.nextElementSibling.value){
+			item.completed = !item.completed;
+		}
+	});
+	localStorage.setItem("listItems", JSON.stringify(taskItems));
+	event.nextElementSibling.classList.toggle("completed")
+}
+
+export default taskCompleted
